input called correctly, comma problem solved, input may= double
1) It is necessary to print to a console point 7, 2) Export should contain state signs, that dont meet the condition 3) The "2)" also applies to point 7, 4) better solution of handling the exception while reading input file, 5) optionally simplify convertToDouble

+++ VAT/src/com/engeto/VAT/Reading.java
@@ -19,11 +19,11 @@ public class Reading {
      * @return Integer read or 0 if invalid input entered
      */
 
-    public static int safeReadInt() {
-        int result = 20;
+    public static double safeReadInt() {
+        double result = 20;
         String inputText = getScanner().nextLine();
         try {
-            result = Integer.parseInt(inputText);
+            result = Double.parseDouble(inputText);
             if (result == 0){result = 20;}
         } catch (NumberFormatException ex) {
             System.err.println("You have entered nothing or ZERO. Default value has been set to 20.");
